# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"MindCraft API","description":"마인드 맵 기반 자기소개서 작성 서비스 API 문서","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"JWT":[]}],"tags":[{"name":"CoverLetter","description":"자기소개서 마스터 관련 API"},{"name":"User","description":"회원 정보 관련 API"},{"name":"CoverLetterSection","description":"자소서 문항 관련 API"},{"name":"Health","description":"EC2 헬스체크용"},{"name":"Token","description":"Access Token 재발급 관련 API"},{"name":"MindMap","description":"마인드맵 관련 API"}],"paths":{"/mindmaps":{"get":{"tags":["MindMap"],"summary":"마인드맵 조회 및 생성","description":"로그인한 사용자의 마인드맵을 조회한다. 없는 경우 루트 노드를 생성해서 반환한다.","operationId":"getMindMap","responses":{"200":{"description":"조회(혹은 생성) 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MindMapResponseDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"put":{"tags":["MindMap"],"summary":"마인드맵 수정사항 저장","description":"로그인한 사용자의 마인드맵을 저장한다.","operationId":"saveMindMap","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MindMapSaveDto"}}},"required":true},"responses":{"200":{"description":"저장 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MindMapResponseDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음, 유효하지 않은 마인드맵, 또는 본인의 마인드맵이 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/coverletters/{id}":{"get":{"tags":["CoverLetter"],"summary":"자소서 마스터 상세 조회","description":"자소서 마스터 id로 해당 자소서의 상세 정보(회사 정보, 문항 목록 등)를 조회한다.","operationId":"getDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CoverLetterDetailDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"본인의 자소서 마스터가 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터 id를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"put":{"tags":["CoverLetter"],"summary":"자소서 마스터 수정","description":"자소서 마스터 id를 통해 자소서 마스터의 정보(회사 이름, 인재상, 직무 설명)를 수정한다.","operationId":"modifyCoverLetter","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoverLetterRequestDto"}}},"required":true},"responses":{"200":{"description":"수정 성공","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"본인의 자소서 마스터가 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터 id를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["CoverLetter"],"summary":"자소서 마스터 삭제","description":"자소서 마스터 id를 통해 자소서 마스터 한 개를 삭제한다.","operationId":"deleteCoverLetter","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"삭제 성공","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"본인의 자소서 마스터가 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터 id를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/coverletters/{coverletterId}/sections/{sectionId}":{"put":{"tags":["CoverLetterSection"],"summary":"자소서 문항 수정","description":"문항 id로 해당 문항의 내용(질문, 답변 등)을 수정한다.","operationId":"update","parameters":[{"name":"coverletterId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sectionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoverLetterSectionDto"}}},"required":true},"responses":{"200":{"description":"수정 성공","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"문항을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["CoverLetterSection"],"summary":"자소서 문항 삭제","description":"문항 id로 해당 문항 한 개를 삭제한다.","operationId":"delete","parameters":[{"name":"coverletterId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sectionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"삭제 성공","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"문항을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/mindmaps/keywords":{"post":{"tags":["MindMap"],"summary":"마인드 노드의 키워드 추출","description":"작성중인 마인드맵 노드에서 AI가 판단한 키워드를 추출한다.","operationId":"extractKeywords","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KeywordExtractRequestDto"}}},"required":true},"responses":{"200":{"description":"추출 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AiKeywordResponse"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"AI 키워드 추출 실패(AI 서버 오류/통신 오류/응답 파싱 실패)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/coverletters":{"get":{"tags":["CoverLetter"],"summary":"자소서 마스터 조회","description":"로그인한 본인의 모든 자소서 마스터를 조회한다.","operationId":"getCoverLetters","responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CoverLetterSummaryDto"}}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"tags":["CoverLetter"],"summary":"자소서 마스터 생성","description":"자소서를 작성할 회사에 대한 정보를 작성하는 자소서 마스터를 생성한다.","operationId":"createCoverLetter","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoverLetterRequestDto"}}},"required":true},"responses":{"201":{"description":"생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CoverLetterSummaryDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음 또는 유효하지 않은 마인드맵","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/coverletters/{coverletterId}/sections":{"get":{"tags":["CoverLetterSection"],"summary":"자소서 문항 목록 조회","description":"자소서 마스터 id로 해당 자소서에 속한 문항(질문/답변) 목록을 조회한다.","operationId":"getList","parameters":[{"name":"coverletterId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CoverLetterSectionDto"}}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"tags":["CoverLetterSection"],"summary":"자소서 문항 생성","description":"자소서 마스터 id에 새로운 문항(질문)을 추가하고, AI가 생성한 답변을 함께 저장한다.","operationId":"create","parameters":[{"name":"coverletterId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoverLetterSectionDto"}}},"required":true},"responses":{"201":{"description":"생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CoverLetterSectionDto"}}}},"400":{"description":"질문 또는 참조 노드(source_node)가 비어 있음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"AI 답변 생성 실패(AI 서버 오류/통신 오류/응답 파싱 실패)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/auth/register":{"post":{"tags":["User"],"summary":"회원 가입","description":"이메일과 비밀번호로 회원가입한다.","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserJoinDto"}}},"required":true},"responses":{"201":{"description":"가입 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"400":{"description":"이미 가입된 이메일","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/users/me":{"get":{"tags":["User"],"summary":"내 정보 조회","description":"로그인한 사용자 본인의 정보를 조회한다.","operationId":"getMyInfo","responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["User"],"summary":"회원 탈퇴","description":"로그인한 사용자 본인이 회원 탈퇴한다.","operationId":"deleteUser","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"탈퇴 성공","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료) 또는 비밀번호가 일치하지 않음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"patch":{"tags":["User"],"summary":"내 정보 수정","description":"로그인한 사용자 본인의 정보(이름이나 비밀번호)를 수정한다.","operationId":"updateMyInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}},"required":true},"responses":{"200":{"description":"수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료) 또는 현재 비밀번호가 일치하지 않음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/users/refresh":{"get":{"tags":["Token"],"summary":"Access Token 재발급","description":"만료된 Access Token을 Refresh Token으로 검증하여 새로운 Access Token을 재발급한다. Refresh Token의 남은 유효기간이 1시간 미만이면 Refresh Token도 함께 재발급한다.","operationId":"refresh","parameters":[{"name":"Authorization","in":"header","description":"Access Token, 'Bearer {token}' 형식","required":true,"schema":{"type":"string"}},{"name":"refreshToken","in":"query","description":"Refresh Token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"재발급 성공 (혹은 기존 Access Token이 아직 유효한 경우 그대로 반환)","content":{"*/*":{"schema":{"type":"string","example":{"accessToken":"(JWT 문자열)","refreshToken":"(JWT 문자열)"}}}}},"401":{"description":"Refresh Token 없음, Authorization 헤더 형식 오류(Bearer 형식 아님), 또는 Refresh Token까지 만료됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"EC2 대상 그룹 헬스 체크 목적","description":"애플리케이션 헬스 체크","operationId":"healthCheck","responses":{"200":{"description":"Healthy","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/coverletters/{id}/export/pdf":{"get":{"tags":["CoverLetter"],"summary":"자소서 내용 PDF 다운","description":"자소서 마스터 id를 통해 해당 자소서 마스터에 포함된 자소서 문항과 내용들을 PDF로 제공한다.","operationId":"exportPdf","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"PDF 제작 성공","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"본인의 자소서 마스터가 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터 id를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/coverletters/{id}/export/docx":{"get":{"tags":["CoverLetter"],"summary":"자소서 내용 DOCX 다운","description":"자소서 마스터 id를 통해 해당 자소서 마스터에 포함된 자소서 문항과 내용들을 DOCX로 제공한다.","operationId":"exportDocx","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"DOCX 제작 성공","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}},"401":{"description":"인증되지 않은 요청(토큰 없음/만료)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"본인의 자소서 마스터가 아님","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"자소서 마스터 id를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"MindMapSaveDto":{"type":"object","properties":{"mindMapId":{"type":"integer","format":"int64"},"title":{"type":"string"},"nodes":{"type":"string"}}},"MindMapResponseDto":{"type":"object","properties":{"mindMapId":{"type":"integer","format":"int64"},"title":{"type":"string"},"nodes":{"type":"string"}}},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"CoverLetterRequestDto":{"type":"object","properties":{"title":{"type":"string"},"mindmap_id":{"type":"integer","format":"int64"},"company_name":{"type":"string","minLength":1},"company_ideal":{"type":"string","minLength":1},"job_description":{"type":"string","minLength":1}},"required":["company_ideal","company_name","job_description","mindmap_id"]},"CoverLetterSectionDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"coverLetterId":{"type":"integer","format":"int64"},"question":{"type":"string"},"answer":{"type":"string"},"sourceNode":{"type":"array","items":{"$ref":"#/components/schemas/ReactNode"}},"writingStyle":{"type":"string"},"maxChars":{"type":"integer","format":"int32"},"allowCreativity":{"type":"boolean"},"selectedNodeIds":{"type":"array","items":{"type":"string"}},"contextNodeIds":{"type":"array","items":{"type":"string"}}}},"ReactNode":{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/ReactNodeData"}}},"ReactNodeData":{"type":"object","properties":{"label":{"type":"string"},"parentId":{"type":"string"},"depth":{"type":"integer","format":"int32"}}},"KeywordExtractRequestDto":{"type":"object","properties":{"nodes":{"type":"array","items":{"$ref":"#/components/schemas/ReactNode"}}}},"AiKeywordResponse":{"type":"object","properties":{"nodes":{"type":"array","items":{"$ref":"#/components/schemas/NodeKeywords"}}}},"NodeKeywords":{"type":"object","properties":{"id":{"type":"string"},"keywords":{"type":"array","items":{"type":"string"}}}},"CoverLetterSummaryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"mindmap_id":{"type":"integer","format":"int64"},"company_name":{"type":"string"},"company_ideal":{"type":"string"},"job_description":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"UserJoinDto":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","name","password"]},"UserResponseDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"provider":{"type":"string","enum":["LOCAL","KAKAO","NAVER","GOOGLE"]}}},"UserUpdateDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"type":"string"},"currentPassword":{"type":"string"}}},"CoverLetterDetailDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"sections":{"type":"array","items":{"$ref":"#/components/schemas/CoverLetterSectionDto"}},"mindmap_id":{"type":"integer","format":"int64"},"company_name":{"type":"string"},"company_ideal":{"type":"string"},"job_description":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}}},"securitySchemes":{"JWT":{"type":"http","name":"JWT","scheme":"bearer","bearerFormat":"JWT"}}}}